refactor(session): introduce session output structure for context window
Updates session management across the agent to use a unified SessionOutput struct
to encapsulate session metadata and runtime context information like the
model context window.

This involves updating server endpoints, client interfaces, and session
models to correctly handle and expose context window data.

+++ pkg/agent/server/types.go
@@ -60,11 +60,11 @@ type ChatResponse struct {
 	Reply            string                  `json:"reply"`
 	Model            string                  `json:"model"`
 	Name             string                  `json:"name,omitempty"`
-	ContextWindow    int                     `json:"ctx_window"`
+	ContextWindow    uint64                  `json:"ctx_window"`
 	ReasoningContent string                  `json:"reasoning_content,omitempty"`
 	Pending          *engine.ApprovalRequest `json:"pending,omitempty"`
 	Usage            *contracts.TokenUsage   `json:"usage,omitempty"`
-	ContextTokens    int                     `json:"context_tokens"`
+	ContextTokens    uint64                  `json:"context_tokens"`
 }
 
 type ChatDeltaEvent struct {
@@ -91,10 +91,10 @@ type ChatDoneEvent struct {
 	ReasoningContent string                  `json:"reasoning_content"`
 	Model            string                  `json:"model"`
 	Name             string                  `json:"name,omitempty"`
-	ContextWindow    int                     `json:"ctx_window"`
+	ContextWindow    uint64                  `json:"ctx_window"`
 	Pending          *engine.ApprovalRequest `json:"pending,omitempty"`
 	Usage            *contracts.TokenUsage   `json:"usage,omitempty"`
-	ContextTokens    int                     `json:"context_tokens"`
+	ContextTokens    uint64                  `json:"context_tokens"`
 }
 
 type ChatErrorEvent struct {
